Move to `scala.testing`
